move config builder, train pipeline,  base callback, etc. to helper.py to clean the codes
