+++ tests/torch/sparsity/movement/test_algo.py
@@ -10,190 +10,28 @@ import onnx
 import pytest
 import torch
 import torch.nn as nn
-from datasets import load_dataset
-from nncf import NNCFConfig
-from nncf.api.compression import CompressionAlgorithmController
 from nncf.common.sparsity.statistics import MovementSparsityStatistics
 from nncf.common.utils.helpers import matches_any, should_consider_scope
-from nncf.torch import create_compressed_model, register_default_init_args
-from nncf.torch.layer_utils import COMPRESSION_MODULES, CompressionParameter
-from nncf.torch.layers import NNCF_MODULES_MAP, NNCFLinear
+from nncf.torch import create_compressed_model
+from nncf.torch.layer_utils import CompressionParameter
+from nncf.torch.layers import NNCFLinear
 from nncf.torch.module_operations import UpdateWeightAndBias
 from nncf.torch.nncf_network import NNCFNetwork
-from nncf.torch.sparsity.functions import apply_binary_mask
 from nncf.torch.sparsity.layers import BinaryMask
 from nncf.torch.sparsity.movement.algo import (ImportanceLoss,
                                                MovementSparsifier,
                                                MovementSparsityController,
                                                SparseConfig, SparseStructure)
-from nncf.torch.utils import get_model_device
 from onnx import numpy_helper
 from pytest import approx
-from tests.torch.helpers import (BasicConvTestModel, MockModel,
-                                 check_correct_nncf_modules_replacement,
-                                 create_compressed_model_and_algo_for_test,
-                                 get_all_inputs_for_graph_node,
-                                 get_empty_config, get_nodes_by_type)
-from transformers import (AutoModelForSequenceClassification, AutoTokenizer,
-                          Trainer, TrainingArguments)
-from transformers.trainer_callback import (TrainerCallback, TrainerControl,
-                                           TrainerState)
-
-print('hello?')
-MODEL_NAME_HF = "google/bert_uncased_L-2_H-128_A-2"
+from tests.torch.sparsity.movement.helpers import (BaseCallback, ConfigBuilder,
+                                                   bert_tiny_torch_model,
+                                                   run_movement_pipeline)
+from transformers import TrainingArguments
+from transformers.trainer_callback import TrainerControl, TrainerState
 
 
-class MovementSparsityConfigBuilder:
-
-    def __init__(self) -> None:
-        self._config_dict = {
-            "input_info": [
-                {"sample_size": [1, 256], "type": "long", "keyword": "input_ids"},
-                {"sample_size": [1, 256], "type": "long", "keyword": "token_type_ids"},
-                {"sample_size": [1, 256], "type": "long", "keyword": "position_ids"},
-                {"sample_size": [1, 256], "type": "long", "keyword": "attention_mask"},
-            ],
-            "compression":
-                {
-                    "algorithm": "movement_sparsity",
-                    "params": {
-                        "schedule": "threshold_polynomial_decay",
-                        "power": 3,
-                        "init_importance_threshold": -0.1,
-                        "final_importance_threshold": 0.0,
-                        "warmup_start_epoch": 1,
-                        "warmup_end_epoch": 3,
-                        "steps_per_epoch": 128 // 32,
-                        "importance_regularization_factor": 0.02,
-                        "update_per_optimizer_step": True,
-                    },
-                    "sparse_structure_by_scopes": [
-                        ["block", [16, 16], "{re}.*attention*"],
-                        ["per_dim", [0], "{re}.*BertIntermediate.*"],
-                        ["per_dim", [1], "{re}.*BertOutput.*"],
-                    ],
-                    "ignored_scopes": [
-                        "{re}embedding", "{re}pooler", "{re}classifier"],
-            },
-        }
-
-    def __call__(self, *args, **kwargs):
-        return self
-
-    def warmup_range(self, a=None, b=None):
-        if a is not None:
-            self._config_dict['compression']['params']['warmup_start_epoch'] = a
-        if b is not None:
-            self._config_dict['compression']['params']['warmup_end_epoch'] = b
-        return self
-
-    def input_info(self, value):
-        self._config_dict['input_info'] = value
-        return self
-
-    def sparse_structure_by_scopes(self, value):
-        self._config_dict['compression']['sparse_structure_by_scopes'] = value
-        return self
-
-    def ignored_scopes(self, value):
-        self._config_dict['compression']['ignored_scopes'] = value
-        return self
-
-    def build(self):
-        return NNCFConfig.from_dict(self._config_dict)
-
-    def get_ignored_scopes(self):
-        return self._config_dict['compression']['ignored_scopes'].copy()
-
-    def get_sparse_structure_by_scopes(self):
-        return self._config_dict['compression']['sparse_structure_by_scopes'].copy()
-
-    def get_warmup_range(self):
-        return (self._config_dict['compression']['params']['warmup_start_epoch'],
-                self._config_dict['compression']['params']['warmup_end_epoch'])
-
-    def get_final_importance_threshold(self):
-        return self._config_dict['compression']['params']['final_importance_threshold']
-
-    def get_importance_regularization_factor(self):
-        return self._config_dict['compression']['params']['importance_regularization_factor']
-
-# @pytest.fixture
-
-
-def yelp_dataset():
-    tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME_HF)
-    dataset = load_dataset("yelp_review_full")
-    max_length = 128
-    n_train_samples, n_val_samples = 128, 128
-
-    def tokenize_fn(examples):
-        row = tokenizer(examples["text"], padding="max_length", truncation=True, max_length=max_length)
-        row['position_ids'] = list(range(max_length))
-        return row
-
-    train_dataset = dataset["train"].shuffle(seed=42).select(range(n_train_samples)).map(tokenize_fn)
-    test_dataset = dataset["test"].shuffle(seed=42).select(range(n_val_samples)).map(tokenize_fn)
-    return train_dataset, test_dataset
-
-
-# @pytest.fixture
-def bert_tiny_torch_model():
-    return AutoModelForSequenceClassification.from_pretrained(MODEL_NAME_HF, num_labels=5)
-
-
-class MovementSparsityTrainer(Trainer):
-    def __init__(self, compression_ctrl, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self.compression_ctrl = compression_ctrl
-
-    def compute_loss(self, model, inputs, return_outputs=False):
-        loss_main, outputs = super().compute_loss(model, inputs, return_outputs=True)
-        loss_compress = self.compression_ctrl.loss()
-        loss = loss_main + loss_compress
-        return (loss, outputs) if return_outputs else loss
-
-
-class MovementSparsityCallback(TrainerCallback):
-    def __init__(self, compression_ctrl: CompressionAlgorithmController) -> None:
-        self.compression_ctrl = compression_ctrl
-        self._compress_log = dict()
-
-    def get_compress_log(self):
-        return self._compress_log.copy()
-
-    def get_train_log(self):
-        return self._train_log.copy()
-
-    def on_train_end(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
-        self._train_log = state.log_history
-
-    def on_step_begin(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
-        self.compression_ctrl.scheduler.step()
-
-    def on_step_end(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
-        movement_ctrl_statistics = self.compression_ctrl.statistics().movement_sparsity
-        info = dict(
-            step=state.global_step,
-            epoch=state.epoch,
-            importance_regularization_factor=movement_ctrl_statistics.importance_regularization_factor,
-            importance_threshold=movement_ctrl_statistics.importance_threshold,
-            relative_sparsity=movement_ctrl_statistics.model_statistics.sparsity_level_for_layers,
-        )
-        self._compress_log[float(state.epoch)] = info
-
-    def on_epoch_begin(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
-        self.compression_ctrl.scheduler.epoch_step()
-        mvmt_ctrl = self.compression_ctrl  # TODO: mutliple child ctrls
-        # should move into the scheduler codes
-        if self.compression_ctrl.scheduler.get_state()["current_epoch"] == mvmt_ctrl.scheduler.warmup_end_epoch:
-            print('Fill stage')
-            mvmt_ctrl.reset_independent_structured_mask()
-            mvmt_ctrl.resolve_structured_mask()
-            mvmt_ctrl.populate_structured_mask()
-
-
-def assert_sparsified_layer_mode(sparsifier: MovementSparsifier, module: nn.Linear, mode, grid_size):
+def check_sparsified_layer_mode(sparsifier: MovementSparsifier, module: NNCFLinear, mode, grid_size):
     weight_shape = module.weight.shape
     assert isinstance(sparsifier._weight_importance, CompressionParameter)
     if mode == SparseStructure.BLOCK:
@@ -210,28 +48,27 @@ def assert_sparsified_layer_mode(sparsifier: MovementSparsifier, module: nn.Line
         assert torch.allclose(sparsifier._bias_importance, ref_bias_importance)
 
 
-@pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder(),  # mixed mode
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([]),  # implicit all fine
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([["fine", [1, 1], "{re}.*"]]),  # explicit all fine
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([["block", [8, 8], "{re}.*"]]),  # all block
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([["per_dim", [0], "{re}.*"]]),  # all per_dim
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([["per_dim", [1], "{re}.*"]]),  # all per_dim
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([["block", [16, 16], "{re}.*query.*"]]),  # mixed of explicit and implicit
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([["per_dim", [0], "{re}.*BertIntermediate.*"]]),
-    MovementSparsityConfigBuilder().sparse_structure_by_scopes([["per_dim", [0], "{re}.*BertIntermediate.*"],
-                                                                ["block", [4, 4], "{re}.*attention.*"]]),
+@ pytest.mark.parametrize('nncf_config_builder', [
+    ConfigBuilder(),  # mixed mode
+    ConfigBuilder(sparse_structure_by_scopes=[]),  # implicit all fine
+    ConfigBuilder(sparse_structure_by_scopes=[["fine", [1, 1], "{re}.*"]]),  # explicit all fine
+    ConfigBuilder(sparse_structure_by_scopes=[["block", [8, 8], "{re}.*"]]),  # all block
+    ConfigBuilder(sparse_structure_by_scopes=[["per_dim", [0], "{re}.*"]]),  # all per_dim
+    ConfigBuilder(sparse_structure_by_scopes=[["per_dim", [1], "{re}.*"]]),  # all per_dim
+    ConfigBuilder(sparse_structure_by_scopes=[["block", [16, 16], "{re}.*query.*"]]),  # mixed of explicit and implicit
+    ConfigBuilder(sparse_structure_by_scopes=[["per_dim", [0], "{re}.*BertIntermediate.*"]]),
+    ConfigBuilder(sparse_structure_by_scopes=[["per_dim", [0], "{re}.*BertIntermediate.*"],
+                                              ["block", [4, 4], "{re}.*attention.*"]]),
     # MovementSparsityConfigBuilder().sparse_structure_by_scopes([["block", [5, 5, 5], "{re}.*"]]),  # wrong config
     # MovementSparsityConfigBuilder().sparse_structure_by_scopes([["block", [5, 5], "{re}.*"]]),  # wrong config
 ])
 def test_can_create_movement_sparsity_layers(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder.build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
     assert isinstance(compression_ctrl, MovementSparsityController)
 
     for scope, module in compressed_model.get_nncf_modules().items():
-        if not should_consider_scope(str(scope), nncf_config_builder.get_ignored_scopes()):
+        if not should_consider_scope(str(scope), nncf_config_builder.get('ignored_scopes')):
             if hasattr(module, 'pre_ops'):
                 for op in module.pre_ops.values():
                     assert (not isinstance(op, UpdateWeightAndBias)) or (not isinstance(op.operand, MovementSparsifier))
@@ -241,76 +78,33 @@ def test_can_create_movement_sparsity_layers(tmp_path, nncf_config_builder):
                 if isinstance(op, UpdateWeightAndBias) and isinstance(op.operand, MovementSparsifier):
                     count_movement_op += 1
                     sparsifier = op.operand
-                    for mode, grid_size, expression in nncf_config_builder.get_sparse_structure_by_scopes():
+                    for mode, grid_size, expression in nncf_config_builder.get('sparse_structure_by_scopes'):
                         if matches_any(str(scope), expression):
-                            assert_sparsified_layer_mode(sparsifier, module, mode, grid_size)
+                            check_sparsified_layer_mode(sparsifier, module, mode, grid_size)
                             break  # only test the first matched expression. Need tests to confirm only one matched expression matched for each layer.
                     else:
-                        assert_sparsified_layer_mode(sparsifier, module, SparseStructure.FINE, (1, 1))
+                        check_sparsified_layer_mode(sparsifier, module, SparseStructure.FINE, (1, 1))
             assert count_movement_op == 1
 
 
-def run_movement_pipeline(tmp_path, compression_ctrl, compressed_model, callbacks: Optional[List[TrainerCallback]] = None, **kwargs):  # temporarily remove fixture for model & dataset
-    train_dataset, eval_dataset = yelp_dataset()
-    args = dict(
-        output_dir=tmp_path / "test_trainer",
-        label_names=["labels"],
-        evaluation_strategy="epoch",
-        per_device_train_batch_size=32,
-        per_device_eval_batch_size=32,
-        num_train_epochs=8,
-        learning_rate=1e-3,
-        optim="adamw_torch",
-        remove_unused_columns=False,
-        report_to=None,
-        disable_tqdm=True,
-        no_cuda=True,  # TODO: where to set cuda devices for cuda training?
-    )
-    args.update(kwargs)
-    training_args = TrainingArguments(**args)
-
-    def compute_metrics(eval_pred):
-        logits, labels = eval_pred
-        predictions = np.argmax(logits, axis=-1)
-        return dict(acc=(predictions == labels).mean())
-
-    if callbacks is None:
-        callbacks = [MovementSparsityCallback(compression_ctrl)]
-    trainer = MovementSparsityTrainer(
-        model=compressed_model,
-        args=training_args,
-        compression_ctrl=compression_ctrl,
-        callbacks=callbacks,
-        train_dataset=train_dataset,
-        eval_dataset=eval_dataset,
-        compute_metrics=compute_metrics,
-    )
-
-    train_result = trainer.train()
-    eval_result = trainer.evaluate()
-    return train_result, eval_result, callbacks
-
-
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder(),
-    MovementSparsityConfigBuilder(),  # TODO: without stuctured_masking
-    MovementSparsityConfigBuilder().warmup_range(0, 1),
-    MovementSparsityConfigBuilder().warmup_range(2, 10),
-    MovementSparsityConfigBuilder().warmup_range(1, 1),
-    MovementSparsityConfigBuilder().warmup_range(-1, -1),  # no warm_up?
-    MovementSparsityConfigBuilder().warmup_range(2, 100)]
-)
+    ConfigBuilder(),  # TODO: without stuctured_masking
+    ConfigBuilder(warmup_start_epoch=0, warmup_end_epoch=1),
+    ConfigBuilder(warmup_start_epoch=2, warmup_end_epoch=5),
+    ConfigBuilder(warmup_start_epoch=1, warmup_end_epoch=1),
+    ConfigBuilder(warmup_start_epoch=-1, warmup_end_epoch=-1),  # no warm_up?
+])
 def test_can_run_full_pipeline(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
-    callback = MovementSparsityCallback(compression_ctrl)
+    callback = BaseCallback(compression_ctrl)
     train_log, eval_result, _ = run_movement_pipeline(tmp_path, compression_ctrl, compressed_model, [callback])
 
     # check
-    warmup_a, warmup_b = nncf_config_builder().get_warmup_range()
-    importance_regularization_factor = nncf_config_builder().get_importance_regularization_factor()
-    final_importance_threshold = nncf_config_builder().get_final_importance_threshold()
+    warmup_a = nncf_config_builder.get('warmup_start_epoch')
+    warmup_b = nncf_config_builder.get('warmup_end_epoch')
+    importance_regularization_factor = nncf_config_builder.get('importance_regularization_factor')
+    final_importance_threshold = nncf_config_builder.get('importance_threshold')
     for epoch, log in callback.get_compress_log().items():
         # TODO: the <= and > may be a bit confusing here.
         # Assume one epoch has 4 steps, and warmup_range is [1, 2]:
@@ -329,14 +123,13 @@ def test_can_run_full_pipeline(tmp_path, nncf_config_builder):
 
 
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder()]
-)
+    ConfigBuilder(),
+])
 def test_importance_score_update(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
 
-    class CheckImportanceCallback(MovementSparsityCallback):
+    class CheckImportanceCallback(BaseCallback):
         def on_epoch_end(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
             super().on_step_end(args, state, control, **kwargs)
             for sparse_module in self.compression_ctrl.sparsified_module_info:
@@ -352,14 +145,13 @@ def test_importance_score_update(tmp_path, nncf_config_builder):
 
 
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder(),
+    ConfigBuilder(),
 ])
 def test_compression_loss(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
 
-    class CheckCompressionLossCallback(MovementSparsityCallback):
+    class CheckCompressionLossCallback(BaseCallback):
         def on_step_end(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
             super().on_step_end(args, state, control, **kwargs)
             assert isinstance(self.compression_ctrl.loss, ImportanceLoss)
@@ -380,15 +172,14 @@ def test_compression_loss(tmp_path, nncf_config_builder):
 
 
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder(),
+    ConfigBuilder(),
 ])
 def test_binary_mask(tmp_path, nncf_config_builder):
     # TODO: ref_binary_mask` in this test is incorrect. We have `max_percentile` argument when thresholding.
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
 
-    class CheckBinaryMask(MovementSparsityCallback):
+    class CheckBinaryMaskCallback(BaseCallback):
         def on_epoch_end(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
             super().on_step_end(args, state, control, **kwargs)
             for sparse_module in self.compression_ctrl.sparsified_module_info:
@@ -397,16 +188,15 @@ def test_binary_mask(tmp_path, nncf_config_builder):
                     ref_binary_mask = (sparsifier._weight_importance > self.compression_ctrl.scheduler.current_importance_threshold).float()
                     assert torch.allclose(sparsifier.weight_ctx.binary_mask, sparsifier._expand_importance(ref_binary_mask))  # TODO: add _expand_importance test in unit test.
 
-    run_movement_pipeline(tmp_path, compression_ctrl, compressed_model, [CheckBinaryMask(compression_ctrl)])
+    run_movement_pipeline(tmp_path, compression_ctrl, compressed_model, [CheckBinaryMaskCallback(compression_ctrl)])
 
 
 # This part is not finished. How to check sparisty with fill stage?
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder(),
+    ConfigBuilder(),
 ])
 def test_sparsity_statistics_are_increasing(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
     assert isinstance(compression_ctrl.statistics().movement_sparsity, MovementSparsityStatistics)
     log_dict = defaultdict(list)
@@ -464,11 +254,10 @@ def check_onnx_has_sparsified_param(compressed_model, compression_ctrl, onnx_pat
 
 
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder(),
+    ConfigBuilder(),
 ])
 def test_export_onnx_has_sparsified_param(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
     for m in compression_ctrl.sparsified_module_info:
         torch.nn.init.uniform_(m.operand._weight_importance, a=-1., b=1.)
@@ -483,11 +272,10 @@ def test_export_onnx_has_sparsified_param(tmp_path, nncf_config_builder):
 
 
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder().warmup_range(0, 1),
+    ConfigBuilder(warmup_start_epoch=0, warmup_end_epoch=1),
 ])
 def test_structured_mask_obeys_unstructured(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
     for m in compression_ctrl.sparsified_module_info:
         torch.nn.init.normal_(m.operand._weight_importance)
@@ -500,7 +288,7 @@ def test_structured_mask_obeys_unstructured(tmp_path, nncf_config_builder):
             mask_dict[(m, 'bias')] = m.operand.bias_ctx.binary_mask.data.clone()
         return mask_dict
 
-    class StructuredMaskingCallback(MovementSparsityCallback):
+    class StructuredMaskingCallback(BaseCallback):
         def on_epoch_begin(self, args: TrainingArguments, state: TrainerState, control: TrainerControl, **kwargs):
             if state.epoch != self.compression_ctrl.scheduler.warmup_end_epoch:
                 super().on_epoch_begin(args, state, control, **kwargs)
@@ -520,24 +308,23 @@ def test_structured_mask_obeys_unstructured(tmp_path, nncf_config_builder):
 
 
 @pytest.mark.parametrize('nncf_config_builder', [
-    MovementSparsityConfigBuilder().warmup_range(0, 1),
+    ConfigBuilder(warmup_start_epoch=0, warmup_end_epoch=1),
 ])
 def test_fill_stage_has_fixed_sparsity(tmp_path, nncf_config_builder):
-    nncf_config = nncf_config_builder().build()
-    nncf_config['log_dir'] = tmp_path
+    nncf_config = nncf_config_builder.build(log_dir=tmp_path)
     compression_ctrl, compressed_model = create_compressed_model(bert_tiny_torch_model(), nncf_config)
     for m in compression_ctrl.sparsified_module_info:
         torch.nn.init.normal_(m.operand._weight_importance)
         torch.nn.init.normal_(m.operand._bias_importance)
 
-    callback = MovementSparsityCallback(compression_ctrl)
+    callback = BaseCallback(compression_ctrl)
     run_movement_pipeline(tmp_path, compression_ctrl, compressed_model, [callback], num_train_epochs=2)
     rela_sparsity = [log['relative_sparsity'] for epoch, log in callback.get_compress_log().items()
-                     if epoch > nncf_config_builder.get_warmup_range()[1]]
+                     if epoch > nncf_config_builder.get('warmup_end_epoch')]
     assert all(sparsity == approx(rela_sparsity[0]) for sparsity in rela_sparsity)
 
 
 if __name__ == "__main__":
     # test_can_create_movement_sparsity_layers(MovementSparsityConfigBuilder())
     # test_can_run_full_pipeline(Path('/tmp'), MovementSparsityConfigBuilder)
-    test_structured_mask_obeys_unstructured(Path('/tmp'), MovementSparsityConfigBuilder().warmup_range(0, 1))
+    test_structured_mask_obeys_unstructured(Path('/tmp'), ConfigBuilder(warmup_start_epoch=0, warmup_end_epoch=1))
